Move ThirdPartyEntityKind into api.constants so the expectation becomes that the value is significant
